pyproject.toml
[project]
name = "mtci"
version = "0.1.0"
description = "Metamorphic Testing for CI"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "MTCI" }]
keywords = ["metamorphic", "testing", "mlops", "ci"]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
hf = ["transformers>=4.40"]

[project.scripts]
mtci = "mtci.cli:app"

[dependency-groups]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mtci"]
